Renamed a few module scripts
